Contrail to support 4 byte AS
Contrail only supports 2 byte AS today. There is a requirement to extend
it to use 4 byte AS also.
This is first draft to support the same. Some testcaes are added and
some more may need to be added.

Partial-Bug: #1416551

+++ src/bgp/bgp_config_ifmap.cc
@@ -1790,7 +1790,7 @@ static void BuildPolicyTermConfig(autogen::PolicyTermType cfg_term,
             cfg_term.term_match_condition.extcommunity_list;
     }
 
-    BOOST_FOREACH(uint16_t asn,
+    BOOST_FOREACH(uint32_t asn,
         cfg_term.term_action_list.update.as_path.expand.asn_list) {
         term->action.update.aspath_expand.push_back(asn);
     }
